return true so complete is called once
Since the return value of all sync calls are returned to the caller, we will call complete for each sync call before the actual one unless we return true from here. As far as anything is concerned, it went async. That is all the more you can tell by the function sync'ly returning true.