Added Tiddlyserver.app
I used a script I found here: https://github.com/labor4/ya-appify/blob/master/appify after reading discussion here https://mathiasbynens.be/notes/shell-script-mac-apps to make a simple shell script into a simple app. The .app is just a folder and all it contains is a single script file. The other bits tell it to open the process in a terminal window. The name of the script file needs to be the same as the name of the .app file. This 'app' will always just try to start a node server in the directory it's in (it doesn't know anything about the app). It should also be possible to stuff all of Tiddlyserver inside the app container, if that's ever desirable.