Interface checker
This is a set of functions that are tightly coupled to the TypeScript interface they validate objects against. This allows us to verify that a third party program (such as the preflighter or a modified server.js) does not pass in unexpected values that are not in the TypeScript interface, because my coding style depends on carefully following Typescript interfaces.