Remove tuples!
Tuples are a pain! Never use them for this. Only use them for very little things or for very big things where you are also writing a const enum to access the Tuple values. The pain comes because Tuples cannot be typechecked against each other for length without the length being manually specified, so there is no short and simple method to do this. It must be maintained with the enum and type declarations.